server/requirements.txt
Flask>=2.3.0
Pillow>=10.0.0
numpy>=1.24.0
Werkzeug>=2.3.0